C++ second week
Write a 2-level menu for a user to navigate. Welcome the user to the program and provide the
user with 3 options to choose from. If the user selects option (a), display the number of students
in our course. If the user selects option (b), display another set of options. If the user selects
option (c), exit the program. For option (b), the sub-menu should include (1) Course textbook,
(2) Assignments, and (3) Exit, for exiting the sub-menu