C++ second week
Tic Tac Toe is a game for two players, X and O, who take turns marking spaces on a 3×3 grid.
The objective of the game is to win by placing three of their marks in a horizontal, vertical, or
diagonal row. If neither player succeeds in doing this and all spaces are taken, then the result is a
tie. In this assignment, you are tasked with coding this game.
Your program will prompt a specific player (X or O) to enter the row and column of the space
that they want to mark. The program will evaluate whether the move is valid. If it is valid, the
space on the board will be updated with a mark corresponding to the player’s symbol. If it is
invalid, the player will be asked to enter the row and column again. The program should run
until Player X wins, Player O wins, or when there’s a tie. A tie occurs when the board is full and
neither player has won. The program ends with the announcement of the game’s outcome.